remove busy loop from destination background future when shutdown (#15)
When the proxy is shutting down, once there are no more outbound
connections, the sender side of the resolver channel is dropped. In the
admin background thread, when the destination background future is
notified of the closure, instead of shutting down itself, it just busy
loops. Now, after seeing shutdown, the background future ends as well.

Signed-off-by: Sean McArthur <sean@buoyant.io>