Share a common tls_config_reload metrics object (#46)
The current implementation of TLS config reload telemetry has several
layers: a sensor emits events into a dispatcher that updates metrics.

This can be simplified by sharing a common metrics object between the
metrics registry and config module.

The metrics registry is updated to hold a read-only `tls_config_reload::Fmt`;
and the config module holds a `tls_config_reload::Sensor`.

The `Sensor` type holds a strong reference to an inner metrics
structure and acquires a lock on updates.

The `Fmt` type holds a weak reference to the metrics structure so
that the metrics server can print the state as long as it's actually held
in memory.  If the `Sensor` is dropped (for instance, because TLS has
been administratively disabled), then no metrics will be formatted by
`Fmt`.