Make outbound router honor `l5d-dst-override` header (#173)
# Problem
It is valuable to inject an ingress controller with LD2. As ingress, the host
header tends to be the external host and most controllers just route directly
back to the IP address of an endpoint. Unfortunately, this makes it so that
Linkerd cannot discover the destination (as that works off the host header).

# Solution
The outbound router now recognizes request destinations by preferrentially
choosing the `l5d-dst-override` header if it present. This header takes
precedent over the Host header, original dest, etc.

In addition, both the inbound and outbound routers strip this header. The
outbound router strips it after it has used it to determine the authority.
This is done with the `strip_header` layer that was added to the stack.

# Validation
There are integration tests for this behavior in the discovery tests. There
are tests for http1 and http2. The tests test the following scenarios:
- the outbound router routes based on the header when it is present
- the override header applies to only requests with the override header
- the inbound router never honors the override header
- both routers strip the header before forwarding the request

Fixes linkerd/linkerd2#1998

Co-Authored-By: Eliza Weisman <eliza@buoyant.io>
Signed-off-by: Kevin Leimkuhler <kevin@kleimkuhler.com>