Update dependencies to remove duplicate `rand` versions (#170)
This branch updates the proxy's dependencies to eliminate multiple
incompatible versions of the `rand` crate. 

Previously we depended on both `rand` 0.4.3 and 0.5.1. After this
branch, we depend on `rand` 0.6.3 (the latest) and 0.5.1. However, PR
#169 will also update `trust-dns-resolver` (the only crate depending on
`rand` 0.5.1) to a version that depends on `rand` 0.6.3, so once both
this branch and #169 are merged, we will depend only on `rand` 0.6.3.

Note that this is a fairly large change to `Cargo.toml` --- it was
necessary to update many of the proxy's dependencies in order to
consolidate on one `rand` version. Additionally, I had to push branches
of some of those dependencies in order to update their `rand`
dependency, so it's currently necessary to patch some of the proxy's
dependencies. When those branches merge upstream, the patches can be
removed.

Signed-off-by: Eliza Weisman <eliza@buoyant.io>