Set `l5d-remote-ip` on inbound requests and outbound responses (#200)
## Problem

[Context](https://github.com/linkerd/linkerd2/issues/2310)

## Solution

We now add the `l5d-remote-ip` header to all inbound requests and outbound responses. We also make sure to strip the header before setting so that clients may not set this header.

We were already adding a header in similar style, so in order to deduplicate code, `add_header.rs` holds the generic logic for adding a header to a `http::Request` or `http::Response`. `l5d-client-id` and `l5d-server-id` now also use the same logic in order to add their respective headers.

## Validation

An additional test module was added to `tests/discovery.rs` that tests the following:
- outbound responses strip `l5d-remote-ip`
- inbound requests strip `l5d-remote-ip`
- outbound responses set `l5d-remote-ip` to the endpoint IP
- inbound requests set `l5d-remote-ip` to the source IP

Closes linkerd/linkerd2#2310

Signed-off-by: Kevin Leimkuhler <kevinl@buoyant.io>