Add test that proxy refreshes its identity (#229)
Depends on #227 and #228.

This branch adds a new integration test that asserts that a proxy will
refresh its identity before the specified expiration time.

I also added a `trace` log in the `identity` module that logs the actual
duration we will wait before attempting to refresh. This was useful when
debugging why this test was failing (it was because I neglected to set
the LINKERD2_PROXY_IDENTITY_MIN_REFRESH env var, and it defaulted to 10
seconds). I felt like it was useful to log the actual duration we will
wait as well as the `SystemTime` when the cert expires, but I'm happy to
back that commit off if we don't actually want that log line.

Closes linkerd/linkerd2#2505

Signed-off-by: Eliza Weisman <eliza@buoyant.io>