Remove buffers from endpoint stacks (#239)
Due to the `http::settings::router`, a `buffer` was needed in each
endpoint stack. This meant that the service was always ready, even if
the client were falling over (and reconnecting). In turn, this meant
that the balancer would pick one of these endpoint stacks, because it
was always ready!

This change includes a test of a failing endpoint, that the balancer no
longer assumes it is ready, and has the following functional changes:

- Removed `http::settings::router`, instead the client HTTP settings are
  detected as part of the `DstAddr`. This means that each balancer only
  has endpoints with the same HTTP settings.
- Removed `buffer` layer from inside the endpoint stacks.

Signed-off-by: Sean McArthur <sean@buoyant.io>