Allow disabling tap by setting an env var (#252)
This PR fixes linkerd/linkerd2#2811. Now if
`LINKERD2_PROXY_TAP_DISABLED` is set, the tap is not served at all. The
approach taken is that  the `ProxyParts` is changed so the
`control_listener` is now an `Option` that will be None if tap is
disabled as this control_listener seems to be exclusively used to serve
the tap. Feel free to suggest a better approach. 

Signed-off-by: Zahari Dichev <zaharidichev@gmail.com>